aniage: Move all routes form init file to api.py
